Adopt math for float api

+++ src/api/dm/addon/math/curves/cycloidal.d
@@ -1,14 +1,14 @@
 module api.dm.addon.math.curves.cycloidal;
 
 import api.dm.addon.math.curves.plane_curves : onPointStep;
-import api.math.geom2.vec2 : Vec2d;
+import api.math.geom2.vec2 : Vec2f;
 
 import Math = api.dm.math;
 
 /**
  * Authors: initkfs
  */
-void hypotrochoid(scope bool delegate(Vec2d) onPointIsContinue, float radius1, float theta1, float radius2, float theta2, float dots = 500, float scale = 1.0)
+void hypotrochoid(scope bool delegate(Vec2f) onPointIsContinue, float radius1, float theta1, float radius2, float theta2, float dots = 500, float scale = 1.0)
 {
     assert(onPointIsContinue);
 
@@ -24,14 +24,14 @@ void hypotrochoid(scope bool delegate(Vec2d) onPointIsContinue, float radius1, f
         const y = (radius1 * Math.sin(
                 theta1 * theta) - radius2 * Math.sin(
                 theta2 * theta)) * scale;
-        if (!onPointIsContinue(Vec2d(x, y)))
+        if (!onPointIsContinue(Vec2f(x, y)))
         {
             break;
         }
     }
 }
 
-void cycloid(scope bool delegate(Vec2d) onPointIsContinue, float radius = 10, size_t dots = 100, float step = 0.5)
+void cycloid(scope bool delegate(Vec2f) onPointIsContinue, float radius = 10, size_t dots = 100, float step = 0.5)
 {
     assert(onPointIsContinue);
 
@@ -39,6 +39,6 @@ void cycloid(scope bool delegate(Vec2d) onPointIsContinue, float radius = 10, si
     onPointStep(step, 0, dots, (dt) {
         const x = (radius * dt) - radius * Math.sin(dt);
         const y = radius - radius * Math.cos(dt);
-        return onPointIsContinue(Vec2d(x, y));
+        return onPointIsContinue(Vec2f(x, y));
     });
 }
